creating the audio only if it has not been created before

+++ example/audio.go
@@ -14,22 +14,29 @@ var (
 	audioFilePath = fmt.Sprintf("./temp/%s", fileName)
 )
 
+func AudioGenerated(filePath string)bool{
+	return fileExists(filePath)
+}
+
 func tts() {
-	audio, err := openai.TextToSpeech(client, httpClient, &openai.SpeechRequestBody{
-		Model: "tts-1",
-		Input: "Hello",
-		Voice: openai.SpeechVoices.Onyx,
-	})
-	if err != nil {
-		panic(err)
-	}
-	defer audio.Close()
-	b, err := io.ReadAll(audio)
-	if err != nil {
-		panic(err)
-	}
-	if err := os.WriteFile(audioFilePath, b, os.ModePerm); err != nil {
-		panic(err)
+	fileExists := AudioGenerated(audioFilePath)
+	if !fileExists{
+		audio, err := openai.TextToSpeech(client, httpClient, &openai.SpeechRequestBody{
+			Model: "tts-1",
+			Input: "Hello",
+			Voice: openai.SpeechVoices.Onyx,
+		})
+		if err != nil {
+			panic(err)
+		}
+		defer audio.Close()
+		b, err := io.ReadAll(audio)
+		if err != nil {
+			panic(err)
+		}
+		if err := os.WriteFile(audioFilePath, b, os.ModePerm); err != nil {
+			panic(err)
+		}
 	}
 }
